waiting on further input to finish, must return to this feature
