+++ client/src/components/StretchModal.js
@@ -2,6 +2,7 @@ import React, { useEffect, useState, useContext } from 'react';
 import { AppContext } from '../context/AppContext';
 import axios from 'axios';
 import { Modal, Card, Button, Dropdown, CardColumns } from 'react-bootstrap';
+import IndividualStretchCard from './IndividualStretchCard';
 
 const StretchModal = (props) => {
   const {
@@ -11,20 +12,20 @@ const StretchModal = (props) => {
     setTargetArea
   } = useContext(AppContext);
 
-  const handleClick = async (e) => {
-    e.preventDefault();
-    console.log(targetArea);
-    console.log(stretchNames);
-    try {
-      const response = await axios.get(`/api/stretches/${targetArea}`);
-      const stretches = response.data.filter((bodyPart) => {
-        return bodyPart.stretchName;
-      });
-      console.log(stretches);
-    } catch {
-      console.log('whoppps');
-    }
-  };
+  //   const handleClick = async (e) => {
+  //     e.preventDefault();
+  //     console.log(targetArea);
+  //     console.log(stretchNames);
+  //     try {
+  //       const response = await axios.get(`/api/stretches/${targetArea}`);
+  //       const stretches = response.data.filter((bodyPart) => {
+  //         return bodyPart.stretchName;
+  //       });
+  //       console.log(stretches);
+  //     } catch {
+  //       console.log('whoppps');
+  //     }
+  //   };
 
   return (
     <Modal {...props}>
@@ -44,32 +45,13 @@ const StretchModal = (props) => {
           <Dropdown.Item>Intermediate</Dropdown.Item>
           <Dropdown.Item>Advanced</Dropdown.Item>
         </Dropdown>
-        <CardColumns>
-          <Card>
-            <Card.Img />
-            <Button onClick={handleClick}></Button>
-          </Card>
-          <Card>
-            <Card.Img />
-            <Button></Button>
-          </Card>
-          <Card>
-            <Card.Img />
-            <Button></Button>
-          </Card>
-          <Card>
-            <Card.Img />
-            <Button></Button>
-          </Card>
-          <Card>
-            <Card.Img />
-            <Button></Button>
-          </Card>
-          <Card>
-            <Card.Img />
-            <Button></Button>
-          </Card>
-        </CardColumns>
+        {stretchNames &&
+          stretchNames.map((stretch) => (
+            <IndividualStretchCard
+              key={stretch._id}
+              stretchName={stretch.stretchName}
+            />
+          ))}
       </Modal.Body>
       <Modal.Footer>
         <Button>Cancel</Button>
